+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleServiceHttp.scala
@@ -1,4 +1,5 @@
-package org.broadinstitute.dsde.workbench.google2
+package org.broadinstitute.dsde.workbench
+package google2
 
 import cats.effect.{Concurrent, Resource, Timer}
 import com.google.cloud.Identity