+++ google2/src/test/scala/org/broadinstitute/dsde/workbench/google2/Generators.scala
@@ -1,4 +1,5 @@
-package org.broadinstitute.dsde.workbench.google2
+package org.broadinstitute.dsde.workbench
+package google2
 
 import java.nio.charset.Charset
 import java.time.Instant