+++ metrics/CHANGELOG.md
@@ -4,10 +4,11 @@ This file documents changes to the `workbench-metrics` library, including notes 
 
 ## 0.5
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.5-4c7acd5"`
-
 ### Changed
 - Moved `SamModel`
+- Publish 2.13
+
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-util" % "0.5-65bba14"`
 
 ## 0.4
 - upgrade cats to 1.4.0 and scala to 2.12.7